moved setup to a harness
